+++ src/test/java/DriverManager/DriverManager.java
@@ -1,6 +1,7 @@
 package DriverManager;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
@@ -23,26 +24,23 @@ public class DriverManager {
         wait = new WebDriverWait(driver, Duration.ofSeconds(30));
 
         fluentWait = new FluentWait<>(driver)
-                .withTimeout(Duration.ofSeconds(30))
-                .pollingEvery(Duration.ofSeconds(2));
+                .withTimeout(Duration.ofSeconds(60))
+                .pollingEvery(Duration.ofSeconds(5))
+                .ignoring(NoSuchElementException.class);
     }
 
     public static WebDriver getDriver(){
         return driver;
     }
 
-    public static WebDriverWait getWait(){
-        return wait;
-    }
-
     public static Wait<WebDriver> getFluentWait(){
         return (Wait<WebDriver>) fluentWait;
     };
 
-    public static WebDriverWait getWaitUntilVisibility(WebElement element){
-        WebDriverWait wait = DriverManager.getWait();
-        wait.until(ExpectedConditions.visibilityOf(element));
-        return wait;
+    public static Wait<WebDriver> getWaitUntilVisibility(WebElement element){
+        Wait<WebDriver> fluentWait = DriverManager.getFluentWait();
+        fluentWait.until(ExpectedConditions.visibilityOf(element));
+        return (Wait<WebDriver>) fluentWait;
     }
 
     public static Wait<WebDriver> getWaitUntilTextToBePresent(WebElement element, String text){
