--- src/test/java/Utilidades/importAllPages.java
@@ -1,11 +1,0 @@
-package Utilidades;
-
-import Pages.CartPage;
-import Pages.ProductDetailPage;
-
-public class importAllPages {
-    Pages.HomePage homePage = new Pages.HomePage();
-    Pages.ProductDetailPage productDetailPage = new ProductDetailPage();
-    Pages.CartPage cartPage = new CartPage();
-    Pages.SuccesPage succesPage2 = new Pages.SuccesPage();
-}
