+++ src/test/java/pom/Services/CartPageServices.java
@@ -1,0 +1,16 @@
+package pom.Services;
+
+import pom.Pages.CartPage;
+
+public class CartPageServices {
+    private final CartPage cartPage;
+
+    public CartPageServices(){
+
+        cartPage = new CartPage();
+    }
+
+
+    public void clickOnBtn_proceedToCheckout() {
+        cartPage.getBtn_proceedToCheckout().click(); }
+    }
